requirements.txt
asgiref==3.8.1
bleach==6.1.0
diff-match-patch==20230430
dj-database-url==2.2.0
Django==5.1
django-import-export==4.1.1
django-markdownify==0.9.5
django-template-partials==24.4
djangorestframework==3.15.2
gunicorn==22.0.0
Markdown==3.7
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
tinycss2==1.2.1
typing_extensions==4.12.2
webencodings==0.5.1
whitenoise==6.7.0
